+++ lib/oneview-sdk/resource/api1600/c7000/enclosure.rb
@@ -9,19 +9,14 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
+require_relative '../../api1200/c7000/enclosure'
+
 module OneviewSDK
   module API1600
-    # Module for API1600 C7000
     module C7000
-      # Get resource class that matches the type given
-      # @param [String] type Name of the desired class type
-      # @return [Class] Resource class or nil if not found
-      def self.resource_named(type)
-        OneviewSDK::API1600.resource_named(type, 'C7000')
+      # Enclosure resource implementation for API1600 C7000
+      class Enclosure < OneviewSDK::API1200::C7000::Enclosure
       end
     end
   end
 end
-
-# Load all API-specific resources:
-Dir[File.dirname(__FILE__) + '/c7000/*.rb'].each { |file| require file }