+++ CHANGELOG.md
@@ -1,9 +1,15 @@
-## v5.12.0
-This release adds support to OneView Rest API version 600/800/1000/1200/1600 for the hardware variants C7000 and Synergy to the following features.
+## v5.12.0 (unreleased)
+
+#### Notes
+This release extends support of the SDK to OneView REST API version 1600 (OneView v5.20).
 
 #### Features supported
+- Enclosure
 - Logical Interconnect
 - Logical Interconnect Group
+- Server Hardware
+- Server Hardware Type
+
 
 ## v5.11.0